feat: ✨ auto-configure using files, new command line flags
## Flags
-c allows to specify a config file to load before all others
-l allows to specify the log level (trace, debug, info, warn, error, fatal)
-o allows to specify the output format (md or html) for issues

### Log-Level
Sets the global log level, defaults to info if it can't parse the given value.

### Auto-Configuration
The following files are considered, with the order determining precedence:

```
given config file via -c flag
<working-dir>/.snyk.env
$HOME/.snyk.env
$HOME/.zshrc.local
$HOME/.zshrc
$HOME/.bashrc
$HOME/.profile
/etc/launchd.conf
/etc/profile
/etc/environment
```

Any lines that contain an environment variable in the format
`VARIABLENAME=VARIABLEVALUE` are added automatically to the environment if not already existent. In case of .profile, .zshrc, etc. if a variable is directly exported e.g. via `export VARIABLENAME=VARIABLEVALUE`, it is not loaded. The export would need to be split of and be in its own line, e.g

```
VARIABLENAME=VARIABLEVALUE
export VARIABLENAME
```

The PATH variable is an exception, as it is treated as an aggregate of all PATH variables found in the files and in the environment. Also, the current working directory is automatically added to the path, so a download of the Snyk CLI into the current working directory would yield a found Snyk CLI.

To find the Snyk CLI, the path is automatically searched for a file named `snyk`, and the first path where it is find is added to the environment. It is later used for OSS and IaC.