feat: add tracking of `delta_scan_type`
This change adds tracking of `delta_scan_type` to the analytics event emitted when a scan is done.  It retrieves the `deltaScanType` from the context and, if present, adds it as a string value to the `extension` map which is part of the analytics event parameters.  This allows analysis of scan performance and behavior segmented by the type of delta scan performed.