fix: use pointers for optional LSP settings [IDE-899] (#811)
Previously we had no way to distinguish between the settings object being omitted and being set to "empty" value for the struct.
As such we were ignoring settings objects that were set to the same as the "empty" struct value, as it would have broken / overridden the default LSP settings for clients that did not support the setting.
With this change to using pointers, we can distinguish between a client not sending the settings object (because they don't support it yet) and a client sending the settings object but deliberately setting the values to be the same as the empty struct.