fix: get IAW form working
Made the form only appear when the IAW FF is enabled in the org.
Made the form fields dynamic like in the web UI.
On submit the form calls `ideSubmitIgnoreRequest` with the form details.
Changed the styles used to not be hardcoded to `--vscode` ones.
Added new style for `--button-text-color`. Renamed `--generated-ai-fix-button-background-color` to `--button-background-color`, and `--container-paddding` to `--container-padding`.
General refactoring and issue fixing for styles.