feat: ignore panel [IDE-1069] [IDE-1076] [IDE-1127] (#840)
* feat: ignore panel

* fix: get IAW form working

Made the form only appear when the IAW FF is enabled in the org.
Made the form fields dynamic like in the web UI.
On submit the form calls `ideSubmitIgnoreRequest` with the form details.
Changed the styles used to not be hardcoded to `--vscode` ones.
Added new style for `--button-text-color`. Renamed `--generated-ai-fix-button-background-color` to `--button-background-color`, and `--container-paddding` to `--container-padding`.
General refactoring and issue fixing for styles.

* fix: change how submit ignore passed to IDEs

Now using a template string `${ideSubmitIgnoreRequest}` instead of a function.
Move logic for HTML template injections to always be set.
Renamed ignore form buttons.

---------

Co-authored-by: Ben Durrans <Benjamin.Durrans@snyk.io>