fix: make auth cancelable [IDE-962] (#888)
Previously you had to wait for the auth request to timeout. This change allows the caller to decide to cancel the current request on demand.
Now when the user tries to auth while a previous auth request is waiting on the callback, the previous request will be immediately cancelled. (Without popping up a failure message)
Also made it so we don't try to fetch the auth token if auth fails and not print a debug success-sounding message.