fix: rework environment fetching
Snapshot the environment before scans (fetching of the project env files) and restore it after. This is so environment variables from one project don't mix into another.
Split reading user shell env and reading project env files to different places.
Close env ready channel instead of writing a message to it, so it can be checked in multiple routines.
Fix incorrect handling of JAVA_HOME introduced in the previous commit.