fix: rework environment fetching [IDE-1314] (#969)
Move initial environment setup to one place that runs in a Goroutine to not block, then wait on the env before scanning to prevent a race condition.
Fix a race condition where the LS settings `Path` was causing a cache of the `PATH` before the environment was fully set up, then later restoring to this old value.

Added an option to `config.New()` to allow tests to create a config without having to search all the directories for binaries.
Also, removed `Config.Path()` as it was only used by tests and `os.Getenv("PATH")` is more reliable.

---------

Co-authored-by: Abdelrahman Shawki Hassan <shawki.hassan@snyk.io>