fix: make async analytics calls not write files
The config analytics calls were getting the folder config via a function which was writing the folder configs to a file.
This was causing the storage to be written to while the test had already finished and was tring to clean up the temporary directory.
Also fixed a test bug where the test relied on the workspace folders being returned in a specific order, but since they are stored in a map, they were not.