test: check scan completion before diags (#1062)
In the smoke test we were waiting on diags, when we should be checking the scan status first, so we can fail early if the scan failed.
Also improved the diags for a scan failure in the test.

Updated the comment to explain the `GetCodeApiUrlForFolder` function's logic.

Updated a missed instance of the `SetUpEngineMock` improvements from 3d84914.