fix: analytics respect folder org [IDE-1501] (#1039)
The decision was made for global config changes and global events to be sent to any org for now, since they are not used by any customer reports.
The tests use a mix of fake workspaces and mock workspaces depending on how much real code we want to test (although mostly it was to do with import loops).

Added a method to get the folder config without any writes, which allows for async goroutines that get folder configs to not cause race conditions.