feat: risk score filtering [IDE-1579] (#1081)
The threshold value is stored in the global config. When the risk score FF is set, only issues matching or exceeding this value are displayed.
Also removed some logging, which was causing excessive spam in our logs as it was per issue.
The logging has been replaced with a summary of issues filtered and why.